Update for further COmanage evaluation
Add detailed instructions for how to enable identity validation,
and document that the group name replacement didn't seem to work.
Reorganize a bit to put numeric GIDs right next to the choice of
group management systems.  Replace the open questions section with
a list of open work items on the COmanage side.

Add some more references to how Grouper could be used to generate
GIDs.